requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
beautifulsoup4==4.11.1
black==23.1.0
Brotli==1.0.9
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.2.0
crispy==0.7.4
crispy-bootstrap5==2023.10
cryptography==38.0.4
cssselect2==0.7.0
cycler==0.12.1
Django==5.0
django-bootstrap==0.2.4
django-bootstrap-v5==1.0.11
django-bootstrap5==23.4
django-crispy-forms==2.1
django-uni-form==0.9.0
django-xhtml2pdf==0.0.4
fabio==2023.10.0
fonttools==4.38.0
freetype-py==2.3.0
future==0.18.2
gobject==0.1.0
h5py==3.10.0
hdf5plugin==4.3.0
html5lib==1.1
idna==3.4
kiwisolver==1.4.5
lxml==4.9.1
matplotlib==3.8.2
mod-wsgi==4.9.4
mypy-extensions==0.4.3
numpy==1.26.3
oscrypto==1.3.0
packaging==23.0
pathspec==0.11.0
Pillow==10.1.0
platformdirs==2.6.2
psycopg2==2.9.4
pycairo==1.25.1
pycparser==2.21
pydyf==0.8.0
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
pyparsing==3.1.1
pypdf==3.12.2
PyPDF2==3.0.1
PyPDF3==1.0.6
pyphen==0.13.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==4.0.8
requests==2.28.1
rlPyCairo==0.3.0
silx==1.1.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.9.0
tzdata==2022.5
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.13
weasyprint==60.2
webencodings==0.5.1
xhtml2pdf==0.2.13
zopfli==0.2.2